+++ src/main/java/com/example/project_machimo/auction/service/AuctionService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface AuctionService {
     public AuctionDTO aList(int id);
-    void highestBidUpdate(int amount, int id);
+    void highestBidUpdate(Long amount, int id);
 }
